♻️  Popup 사용 방식 변경에 따른 코드 수정
